requirements.txt
amqp==5.1.1
async-generator==1.10
attrs==22.1.0
billiard==3.6.4.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
outcome==1.2.0
prompt-toolkit==3.0.30
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.40
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
vine==5.0.0
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
